refactored all models,serializers and imports
