+++ backend/.flake8
@@ -10,6 +10,7 @@ exclude =
     .env,
     .idea,
     .venv,
+    user/migrations/*,
     ../frontend/*,
     ./frontend/*,
     ../bot/venv,
@@ -20,4 +21,5 @@ exclude =
     ../info-service/*,
     ../tg-frontend/*,
     ../web-frontend/*
+
