+++ backend/auth_app/models.py
@@ -2,35 +2,6 @@ from django.db import models
 from django.contrib.auth.models import AbstractBaseUser, BaseUserManager
 
 
-class Map(models.Model):
-    map_id = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=255)
-    link = models.CharField(max_length=255, null=True)
-    image_link = models.CharField(max_length=255, null=True)
-
-    def __str__(self):
-        return self.name
-
-
-class GrenadeClass(models.Model):
-    grenade_class_id = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=255)
-    description = models.CharField(max_length=255, null=True, blank=True)
-    price = models.IntegerField(default=0)
-
-    def __str__(self):
-        return self.name
-
-
-class Property(models.Model):
-    property_id = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=255)
-    value = models.CharField(max_length=255, null=True)
-
-    def __str__(self):
-        return self.name
-
-
 class UserManager(BaseUserManager):
     def create_user(self, username, email, password=None):
         if not email:
@@ -58,34 +29,6 @@ class User(AbstractBaseUser):
         return self.username
 
 
-class Lineup(models.Model):
-    grenade_id = models.AutoField(primary_key=True)
-    map_id = models.ForeignKey(Map, on_delete=models.CASCADE)
-    link_to_video = models.CharField(max_length=255, null=True)
-    user_id = models.ForeignKey(User, on_delete=models.CASCADE)
-    created_at = models.DateTimeField(auto_now_add=True)
-    title = models.CharField(max_length=255)
-    description = models.TextField(null=True)
-    is_approved = models.BooleanField(default=False)
-    views = models.IntegerField(default=0)
-    preview_image_link = models.CharField(max_length=255, null=True)
-    grenade_class_id = models.ForeignKey(GrenadeClass, on_delete=models.CASCADE)
-
-    def __str__(self):
-        return self.title
-
-
-class PropertyList(models.Model):
-    property_id = models.ForeignKey(Property, on_delete=models.CASCADE)
-    grenade_id = models.ForeignKey(Lineup, on_delete=models.CASCADE)
-
-    class Meta:
-        unique_together = ("property_id", "grenade_id")
-
-    def __str__(self):
-        return f"{self.property_id} — {self.grenade_id}"
-
-
 class AdminType(models.Model):
     admin_type_id = models.AutoField(primary_key=True)
     is_superuser = models.BooleanField(default=False)
@@ -105,15 +48,3 @@ class Admins(models.Model):
 
     def __str__(self):
         return f"Admin {self.user_id}"
-
-
-class Favorites(models.Model):
-    user_id = models.ForeignKey(User, on_delete=models.CASCADE)
-    grenade_id = models.ForeignKey(Lineup, on_delete=models.CASCADE)
-    created_at = models.DateTimeField(auto_now_add=True)
-
-    class Meta:
-        unique_together = (("user_id", "grenade_id"),)
-
-    def __str__(self):
-        return f"Favorite {self.user_id} - {self.grenade_id}"
