+++ backend/backend/settings.py
@@ -47,8 +47,12 @@ INSTALLED_APPS = [
     "rest_framework_simplejwt",
     "drf_spectacular",
     "auth_app",
-    "lineups",
     "user",
+    "lineups",
+    "maps",
+    "grenade_class",
+    "properties",
+    "favorites",
 ]
 
 REST_FRAMEWORK = {
