+++ backend/grenade_class/views.py
@@ -1,8 +1,8 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from auth_app.models import GrenadeClass
-from auth_app.serializers import GrenadeClassSerializer
+from .models import GrenadeClass
+from .serializers import GrenadeClassSerializer
 from django.shortcuts import get_object_or_404
 from drf_spectacular.utils import extend_schema, OpenApiExample
 from rest_framework.permissions import IsAuthenticated
