+++ backend/maps/views.py
@@ -2,8 +2,8 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from django.shortcuts import get_object_or_404
-from auth_app.models import Map
-from auth_app.serializers import MapSerializer, MapDetailSerializer
+from .models import Map
+from .serializers import MapSerializer, MapDetailSerializer
 from drf_spectacular.utils import extend_schema, OpenApiExample
 from rest_framework.permissions import IsAuthenticated
