+++ backend/properties/property_views.py
@@ -1,8 +1,8 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from auth_app.models import Property
-from auth_app.serializers import PropertySerializer
+from .models import Property
+from .serializers import PropertySerializer
 from django.shortcuts import get_object_or_404
 from drf_spectacular.utils import extend_schema, OpenApiExample
 from rest_framework.permissions import IsAuthenticated
